Aggiunto Swagger e risolti alcuni bug

+++ src/main/java/java13g3/justMeet/controllers/EventController.java
@@ -102,8 +102,11 @@ public class EventController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<String> deleteEventById(@PathVariable("id") Long id) {
-        eventService.deleteEventById(id);
-        return ResponseEntity.ok("L'evento è stato cancellato");
+        if(eventService.retrieveEventById(id).isPresent()){
+           eventService.deleteEventById(id);
+           return ResponseEntity.ok("L'evento è stato cancellato");
+        };
+           return ResponseEntity.notFound().build();
     }
 }
